+++ server/src-lib/Hasura/GraphQL/Transport/WebSocket.hs
@@ -385,9 +385,14 @@ onStart env serverEnv wsConn (StartMsg opId q) = catchAndIgnore $ do
             E.ExecStepDB _remoteHeaders exists ->
               AB.dispatchAnyBackend @BackendTransport exists EB.getRemoteSchemaInfo
             _ -> []
+          actionsInfo = foldl getExecStepActionWithActionInfo [] $ OMap.elems $ OMap.filter (\x -> case x of
+              E.ExecStepAction (_, _) -> True
+              _                       -> False
+              ) queryPlan
+
       -- We ignore the response headers (containing TTL information) because
       -- WebSockets don't support them.
-      (_responseHeaders, cachedValue) <- Tracing.interpTraceT (withExceptT mempty) $ cacheLookup remoteJoins cacheKey
+      (_responseHeaders, cachedValue) <- Tracing.interpTraceT (withExceptT mempty) $ cacheLookup remoteJoins actionsInfo cacheKey
       case cachedValue of
         Just cachedResponseData -> do
           sendSuccResp cachedResponseData $ LQ.LiveQueryMetadata 0
@@ -409,7 +414,7 @@ onStart env serverEnv wsConn (StartMsg opId q) = catchAndIgnore $ do
               return $ ResultsFragment telemTimeIO_DT Telem.Local resp []
             E.ExecStepRemote rsi gqlReq -> do
               runRemoteGQ fieldName userInfo reqHdrs rsi gqlReq
-            E.ExecStepAction actionExecPlan -> do
+            E.ExecStepAction (actionExecPlan, _) -> do
               (time, (r, _)) <- doQErr $ EA.runActionExecution actionExecPlan
               pure $ ResultsFragment time Telem.Empty r []
             E.ExecStepRaw json ->
@@ -458,7 +463,7 @@ onStart env serverEnv wsConn (StartMsg opId q) = catchAndIgnore $ do
                          tx
                          genSql
               return $ ResultsFragment telemTimeIO_DT Telem.Local resp []
-            E.ExecStepAction actionExecPlan -> do
+            E.ExecStepAction (actionExecPlan, _) -> do
               (time, (r, hdrs)) <- doQErr $ EA.runActionExecution actionExecPlan
               pure $ ResultsFragment time Telem.Empty r $ fromMaybe [] hdrs
             E.ExecStepRemote rsi gqlReq -> do
@@ -522,6 +527,10 @@ onStart env serverEnv wsConn (StartMsg opId q) = catchAndIgnore $ do
 
       liftIO $ logOpEv ODStarted (Just requestId)
   where
+    getExecStepActionWithActionInfo acc execStep = case execStep of
+       E.ExecStepAction (_, actionInfo) -> (actionInfo:acc)
+       _                                -> acc
+
     doQErr = withExceptT Right
 
     forWithKey = flip OMap.traverseWithKey
@@ -651,7 +660,6 @@ onStart env serverEnv wsConn (StartMsg opId q) = catchAndIgnore $ do
 
     catchAndIgnore :: ExceptT () m () -> m ()
     catchAndIgnore m = void $ runExceptT m
-
 
 onMessage
   :: ( HasVersion