+++ server/src-lib/Hasura/RQL/Types/Action.hs
@@ -31,6 +31,7 @@ module Hasura.RQL.Types.Action
   , aiOutputObject
   , aiDefinition
   , aiPermissions
+  , aiForwardedClientHeaders
   , aiComment
   , defaultActionTimeoutSecs
   , ActionPermissionInfo(..)
@@ -61,6 +62,9 @@ module Hasura.RQL.Types.Action
   , ActionLogResponse(..)
   , ActionLogResponseMap
   , AsyncActionStatus(..)
+  , ActionsInfo(..)
+  , asiName
+  , asiForwardClientHeaders
   ) where
 
 
@@ -204,11 +208,12 @@ getActionOutputFields =
 
 data ActionInfo
   = ActionInfo
-  { _aiName         :: !ActionName
-  , _aiOutputObject :: !(G.GType, AnnotatedObjectType)
-  , _aiDefinition   :: !ResolvedActionDefinition
-  , _aiPermissions  :: !ActionPermissionMap
-  , _aiComment      :: !(Maybe Text)
+  { _aiName                   :: !ActionName
+  , _aiOutputObject           :: !(G.GType, AnnotatedObjectType)
+  , _aiDefinition             :: !ResolvedActionDefinition
+  , _aiPermissions            :: !ActionPermissionMap
+  , _aiForwardedClientHeaders :: !Bool
+  , _aiComment                :: !(Maybe Text)
   } deriving (Generic)
 instance J.ToJSON ActionInfo where
   toJSON = J.genericToJSON hasuraJSON
@@ -316,8 +321,9 @@ traverseAnnActionExecution f (AnnActionExecution n ot fs p oF dl w h fch sn to s
 
 data AnnActionMutationAsync
   = AnnActionMutationAsync
-  { _aamaName    :: !ActionName
-  , _aamaPayload :: !J.Value -- ^ jsonified input arguments
+  { _aamaName                 :: !ActionName
+  , _aamaForwardClientHeaders :: !Bool
+  , _aamaPayload              :: !J.Value -- ^ jsonified input arguments
   } deriving (Show, Eq)
 
 data AsyncActionQueryFieldG (b :: BackendType) v
@@ -335,21 +341,22 @@ traverseAsyncActionQueryField
 traverseAsyncActionQueryField f = \case
   AsyncTypename t    -> pure $ AsyncTypename t
   AsyncOutput fields -> AsyncOutput <$> traverse (traverse $ traverseAnnField f) fields
-  AsyncId            -> pure $ AsyncId
-  AsyncCreatedAt     -> pure $ AsyncCreatedAt
-  AsyncErrors        -> pure $ AsyncErrors
+  AsyncId            -> pure AsyncId
+  AsyncCreatedAt     -> pure AsyncCreatedAt
+  AsyncErrors        -> pure AsyncErrors
 
 type AsyncActionQueryFieldsG b v = Fields (AsyncActionQueryFieldG b v)
 
 data AnnActionAsyncQuery (b :: BackendType) v
   = AnnActionAsyncQuery
-  { _aaaqName           :: !ActionName
-  , _aaaqActionId       :: !ActionId
-  , _aaaqOutputType     :: !GraphQLType
-  , _aaaqFields         :: !(AsyncActionQueryFieldsG b v)
-  , _aaaqDefinitionList :: ![(Column b, ScalarType b)]
-  , _aaaqStringifyNum   :: !Bool
-  , _aaaqSource         :: !(ActionSourceInfo b)
+  { _aaaqName                 :: !ActionName
+  , _aaaqActionId             :: !ActionId
+  , _aaaqOutputType           :: !GraphQLType
+  , _aaaqFields               :: !(AsyncActionQueryFieldsG b v)
+  , _aaaqDefinitionList       :: ![(Column b, ScalarType b)]
+  , _aaaqStringifyNum         :: !Bool
+  , _aaaqForwardClientHeaders :: !Bool
+  , _aaaqSource               :: !(ActionSourceInfo b)
   }
 
 traverseAnnActionAsyncQuery
@@ -357,8 +364,8 @@ traverseAnnActionAsyncQuery
   => (a -> f b)
   -> AnnActionAsyncQuery backend a
   -> f (AnnActionAsyncQuery backend b)
-traverseAnnActionAsyncQuery f (AnnActionAsyncQuery n aid ot fs dl sn s) =
-  traverse (traverse $ traverseAsyncActionQueryField f) fs <&> \tfs -> AnnActionAsyncQuery n aid ot tfs dl sn s
+traverseAnnActionAsyncQuery f (AnnActionAsyncQuery n aid ot fs dl sn fch s) =
+  traverse (traverse $ traverseAsyncActionQueryField f) fs <&> \tfs -> AnnActionAsyncQuery n aid ot tfs dl sn fch s
 
 data ActionExecContext
   = ActionExecContext
@@ -397,3 +404,11 @@ type ActionLogResponseMap = HashMap ActionId ActionLogResponse
 data AsyncActionStatus
   = AASCompleted !J.Value
   | AASError !QErr
+
+data ActionsInfo
+  = ActionsInfo
+  { _asiName                 :: !ActionName
+  , _asiForwardClientHeaders :: !Bool
+  }
+  deriving (Show, Eq, Generic)
+$(makeLenses ''ActionsInfo)