+++ CHANGELOG.md
@@ -134,6 +134,7 @@ query {
 - server: fix action custom types failing to parse when mutually recursive
 - server: fix MSSQL table name descriptions
 - server: emit `postgres-max-connections-error` when max postgres connections are reached
+- server: disable caching for actions when "forward-client-headers" option is turned on
 - console: allow editing rest endpoints queries and misc ui improvements
 - console: display collection names and queries from all collections in allowlist
 - cli: match ordering of keys in project metadata files with server metadata